+++ src/extensions/core/load3d/Load3dUtils.ts
@@ -1,7 +1,7 @@
 import { t } from '@/i18n'
+import { useToastStore } from '@/platform/updates/common/toastStore'
 import { api } from '@/scripts/api'
 import { app } from '@/scripts/app'
-import { useToastStore } from '@/stores/toastStore'
 
 class Load3dUtils {
   static async uploadTempImage(
