+++ src/platform/updates/common/releaseStore.ts
@@ -3,10 +3,11 @@ import { defineStore } from 'pinia'
 import { computed, ref } from 'vue'
 
 import { useSettingStore } from '@/platform/settings/settingStore'
-import { type ReleaseNote, useReleaseService } from '@/services/releaseService'
 import { useSystemStatsStore } from '@/stores/systemStatsStore'
 import { isElectron } from '@/utils/envUtil'
 import { compareVersions, stringToLocale } from '@/utils/formatUtil'
+
+import { type ReleaseNote, useReleaseService } from './releaseService'
 
 // Store for managing release notes
 export const useReleaseStore = defineStore('release', () => {
