+++ src/platform/updates/common/versionCompatibilityStore.ts
