+++ src/platform/updates/components/ReleaseNotificationToast.vue
@@ -48,9 +48,10 @@
 import { computed, onMounted, ref, watch } from 'vue'
 import { useI18n } from 'vue-i18n'
 
-import type { ReleaseNote } from '@/services/releaseService'
-import { useReleaseStore } from '@/stores/releaseStore'
 import { formatVersionAnchor } from '@/utils/formatUtil'
+
+import type { ReleaseNote } from '../common/releaseService'
+import { useReleaseStore } from '../common/releaseStore'
 
 const { locale } = useI18n()
 const releaseStore = useReleaseStore()
