+++ src/platform/workflow/validation/composables/useWorkflowValidation.ts
@@ -1,7 +1,7 @@
 import type { ISerialisedGraph } from '@/lib/litegraph/src/types/serialisation'
+import { useToastStore } from '@/platform/updates/common/toastStore'
 import type { ComfyWorkflowJSON } from '@/platform/workflow/validation/schemas/workflowSchema'
 import { validateComfyWorkflow } from '@/platform/workflow/validation/schemas/workflowSchema'
-import { useToastStore } from '@/stores/toastStore'
 import { fixBadLinks } from '@/utils/linkFixer'
 
 interface ValidationResult {
